requirements.txt
flask==3.0.2
flask-wtf==1.2.1
spacy==3.8.0
PyPDF2==3.0.1
python-docx==0.8.11
scikit-learn==1.4.0
werkzeug==3.0.1
python-json-logger==2.0.7